+++ hugo-command/main.js
@@ -1732,6 +1732,7 @@ var AboutModal = class extends import_obsidian5.Modal {
       text: "github.com/robotpony/obsidian-plugins",
       href: "https://github.com/robotpony/obsidian-plugins"
     });
+    details.createEl("p", { text: "Made in \u{1F1E8}\u{1F1E6}", cls: "about-made-in" });
   }
   onClose() {
     this.contentEl.empty();